requirements.txt
fastapi==0.101.1
uvicorn==0.23.2
python-dotenv==1.0.0
stripe==6.6.0
flask==2.3.3
black==23.7.0
isort==5.12.0
mypy==1.8.0
pyright==1.1.403  # Use the version installed by `pip install pyright` (check with `pyright --version`)